backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
cryptography==43.0.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
openai==1.54.4
python-magic-bin==0.4.14
gitpython==3.1.43
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
